DX: use anonymous function over concrete classes
