--- tests/Differ/DummyTestSplFileInfo.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of PHP CS Fixer.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *     Dariusz Rumiński <dariusz.ruminski@gmail.com>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
-namespace PhpCsFixer\Tests\Differ;
-
-final class DummyTestSplFileInfo extends \SplFileInfo
-{
-    public function getRealPath(): string
-    {
-        return $this->getFilename();
-    }
-}
